+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.2.6
 - Apply filtering to Command Palette for certain commands
 - Automatically refresh the campaign when 'campaign.json' is modified
+- Handle JSON components specifying templateItem
 
 ## 0.2.5
 - Fix pretty nasty README issue that screwed up most of the documentation